+++ src/editor/contentState/paragraphCtrl.js
@@ -30,6 +30,16 @@ const paragraphCtrl = ContentState => {
     end.type = endBlock.type
     end.block = endBlock
 
+    if (start.type === 'pre' && end.type === 'pre') {
+      const preElement = document.querySelector(`#${start.key}`)
+      const { y } = preElement.getBoundingClientRect()
+      const { line } = start.block.pos
+      const FONT_SIZE = 15 // TODO: in the future font size be be set in perference.
+      const STYLE_LINE_HEIGHT = 1.6 // in stylesheet
+      const LINE_HEIGHT = STYLE_LINE_HEIGHT * FONT_SIZE
+      cursorCoords.y = y + line * LINE_HEIGHT
+    }
+
     return {
       start,
       end,