+++ src/main/actions/view.js
@@ -1,4 +1,26 @@
+import { ipcMain, BrowserWindow } from 'electron'
+import { getMenuItem } from '../utils'
+
+const HASH = {
+  'Source Code Mode': 'sourceCode',
+  'Typewriter Mode': 'typewriter',
+  'Focus Mode': 'focus'
+}
+
 export const view = (win, item, type) => {
   const { checked } = item
   win.webContents.send('AGANI::view', { type, checked })
 }
+
+ipcMain.on('AGANI::ask-for-mode', e => {
+  // format menu
+  const viewMenuItem = getMenuItem('View')
+  const modes = {}
+  viewMenuItem.submenu.items.forEach(item => {
+    if (HASH[item.label]) {
+      modes[HASH[item.label]] = item.checked
+    }
+  })
+  const win = BrowserWindow.fromWebContents(e.sender)
+  win.webContents.send('AGANI::res-for-mode', modes)
+})