+++ src/renderer/components/search.vue
@@ -65,12 +65,32 @@
       <div class="input-wrapper replace-input">
         <input type="text" v-model="replaceValue" placeholder="Replacement">
       </div>
-      <button class="button" @click="replace(false)">
-        All
-      </button>
-      <button class="button" @click="replace(true)">
-        Replace
-      </button>
+      <el-tooltip class="item" 
+        effect="dark"
+        content="Replace All"
+        placement="top"
+        :visible-arrow="false"
+        :open-delay="1000"
+      >
+        <button class="button" @click="replace(false)">
+          <svg class="icon" aria-hidden="true">
+            <use xlink:href="#icon-all-inclusive"></use>
+          </svg>
+        </button>
+      </el-tooltip>
+      <el-tooltip class="item" 
+        effect="dark"
+        content="Replace Single"
+        placement="top"
+        :visible-arrow="false"
+        :open-delay="1000"
+      >
+        <button class="button" @click="replace(true)">
+          <svg class="icon" aria-hidden="true">
+            <use xlink:href="#icon-replace"></use>
+          </svg>
+        </button>
+      </el-tooltip>
     </section>
   </div>
 </template>